[Add] complete file split
