+++ frontend/layouts/PcDetails.vue
@@ -82,6 +82,9 @@ export default {
   components: {
     FriendRequestBtn
   },
+  props: {
+    user: Object
+  },
   data() {
     return {
       tab: null,
@@ -100,27 +103,9 @@ export default {
       resultRequestType: null,
       resultRequestMessage: null
     }
-  },
-  async asyncData({ $axios, params, error }) {
-    const response = await $axios
-      .get(`/api/users/${params.id}`)
-      .catch((err) => {
-        console.log(err)
-        console.log(err.response)
-        return err.response
-      })
-
-    if (!response) {
-      throw new Error('ネットワークに接続できませんでした')
-      // TODO: もう少し何とかする。toastとか出す?
-    } else if (response.status !== 200) {
-      error({ status: response.status, message: response.data.message })
-      return
-    }
-
-    return {
-      user: response.data.data
-    }
   }
 }
 </script>
+<style scoped>
+@import '~/assets/css/pc_details.css';
+</style>
